requirements.txt
typing
openai
pydantic
langchain
python-dotenv
langchain-openai